About: delete and move the content
